fix: Changes as per PR comments
