--- src/main/java/com/autonomic/tmc/auth/AuthenticationCommunicationException.java
@@ -1,36 +1,0 @@
-/*-
- * ‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾
- * tmc-auth
- * ——————————————————————————————————————————————————————————————————————————————
- * Copyright (C) 2016 - 2019 Autonomic, LLC - All rights reserved
- * ——————————————————————————————————————————————————————————————————————————————
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at:
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations
- * under the License
- * ______________________________________________________________________________
- */
-package com.autonomic.tmc.auth;
-
-/**
- * Runtime exception thrown when there is a communication issue while attempting auth. IOException or
- * UnknownHostException being thrown are examples of communication issues.
- */
-@Deprecated
-public class AuthenticationCommunicationException extends RuntimeException {
-
-    public AuthenticationCommunicationException(String msg, Throwable cause) {
-        super(msg, cause);
-    }
-
-    public AuthenticationCommunicationException(String msg) {
-        super(msg);
-    }
-}
