--- src/main/java/com/autonomic/tmc/auth/AuthenticationFailedException.java
@@ -1,33 +1,0 @@
-/*-
- * ‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾
- * tmc-auth
- * ——————————————————————————————————————————————————————————————————————————————
- * Copyright (C) 2016 - 2019 Autonomic, LLC - All rights reserved
- * ——————————————————————————————————————————————————————————————————————————————
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at:
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations
- * under the License
- * ______________________________________________________________________________
- */
-package com.autonomic.tmc.auth;
-
-/**
- * Exception thrown when a response from the token provider has a status code of 400 or 401
- * indicating that the credentials provided have been expressly rejected.
- */
-@Deprecated
-public class AuthenticationFailedException extends RuntimeException {
-
-    public AuthenticationFailedException(String msg) {
-        super(msg);
-    }
-
-}
